+++ projetocampeonato/ClubeBrasileirao.java
@@ -6,7 +6,7 @@ public class ClubeBrasileirao extends Clube{
     private int pontuacao; // Varíavel que armazena a pontuação do clube
     private ArrayList<ClubeBrasileirao> confrontos; // Variável que armazena os clubes que o clube já enfrentou
 
-    public ClubeBrasileirao(String nome, int fundacao, String local, double torcida, int score) {
+    public ClubeBrasileirao(String nome, int fundacao, String local, long torcida, int score) {
         super(nome, fundacao, local, torcida, score);
         setPontuacao(0); // Inicia cada clube com 0 pontos
         confrontos = new ArrayList<ClubeBrasileirao>();
